feat: implement exception handling

+++ src/main/java/dev/anton_kulakov/service/handler/FileResourceHandler.java
@@ -1,6 +1,8 @@
 package dev.anton_kulakov.service.handler;
 
 import dev.anton_kulakov.dto.ResourceInfoDto;
+import dev.anton_kulakov.exception.ResourceAlreadyExistsException;
+import dev.anton_kulakov.exception.ResourceNotFoundException;
 import dev.anton_kulakov.mapper.ResourceMapper;
 import dev.anton_kulakov.service.MinioService;
 import lombok.RequiredArgsConstructor;
@@ -15,16 +17,32 @@ public class FileResourceHandler implements ResourceHandlerInterface {
 
     @Override
     public ResourceInfoDto getInfo(String path) {
+        if (!minioService.isFileExists(path)) {
+            throw new ResourceNotFoundException("The requested file could not be found");
+        }
+
         return resourceMapper.toFileInfoDto(minioService.getStatObject(path));
     }
 
     @Override
     public void delete(String path) {
+        if (!minioService.isFileExists(path)) {
+            throw new ResourceNotFoundException("The requested file could not be found");
+        }
+
         minioService.removeObject(path);
     }
 
     @Override
     public void move(String from, String to) {
+        if (!minioService.isFileExists(from)) {
+            throw new ResourceNotFoundException("The requested file could not be found");
+        }
+
+        if (minioService.isFileExists(to)) {
+            throw new ResourceAlreadyExistsException("The file already exists at the destination path: %s".formatted(to));
+        }
+
         minioService.copy(from, to);
         minioService.removeObject(from);
     }
@@ -36,6 +54,10 @@ public class FileResourceHandler implements ResourceHandlerInterface {
 
     @Override
     public ResourceInfoDto upload(String path, MultipartFile file) {
+        if (minioService.isFileExists(path)) {
+            throw new ResourceAlreadyExistsException("The file already exists at the destination path: %s".formatted(path));
+        }
+
         minioService.upload(path, file);
         String newPath = path + file.getOriginalFilename();
         return getInfo(newPath);
